requirements.txt
google-genai>=0.3.0
streamlit
